chore: move duplicate logic to sdk
chore: move util fns to top
