+++ packages/server/src/api/controllers/workspaceApp.ts
@@ -12,7 +12,6 @@ import {
 } from "@budibase/types"
 import sdk from "../../sdk"
 import { defaultAppNavigator } from "../../constants/definitions"
-import { helpers } from "@budibase/shared-core"
 
 function toWorkspaceAppResponse(
   workspaceApp: WorkspaceApp
@@ -37,57 +36,6 @@ export async function fetch(ctx: Ctx<void, FetchWorkspaceAppResponse>) {
   }
 }
 
-const duplicateScreens = async (originalAppId: string, newAppId: string) => {
-  const screens = await sdk.screens.fetch()
-
-  const appScreens = screens.filter(s => s.workspaceAppId === originalAppId)
-  const newScreens = []
-  for (let i = 0; i < appScreens.length; i++) {
-    const screen = appScreens[i]
-    const createdScreen = await sdk.screens.create({
-      ...{
-        layoutId: screen.layoutId,
-        showNavigation: screen.showNavigation,
-        width: screen.width,
-        routing: screen.routing,
-        props: screen.props,
-        name: screen.name,
-        pluginAdded: screen.pluginAdded,
-        onLoad: screen.onLoad,
-        variant: screen.variant,
-      },
-      workspaceAppId: newAppId,
-    })
-
-    newScreens.push(createdScreen)
-  }
-
-  return newScreens
-}
-
-const createDuplicatedApp = async (workspaceApp: WorkspaceApp) => {
-  const otherApps = await sdk.workspaceApps.fetch()
-
-  const name = helpers.duplicateName(
-    workspaceApp.name,
-    otherApps.map(a => a.name)
-  )
-
-  const duplicatedAppData = {
-    name,
-    url: `/${slugify(name)}`,
-    disabled: true,
-    navigation: workspaceApp.navigation,
-    isDefault: false,
-  }
-
-  const duplicatedApp = await sdk.workspaceApps.create(duplicatedAppData)
-
-  return duplicatedApp
-}
-
-const slugify = (text: string) => text.toLowerCase().replaceAll(" ", "-")
-
 export async function duplicate(
   ctx: Ctx<void, InsertWorkspaceAppResponse, { id: string }>
 ) {
@@ -97,11 +45,7 @@ export async function duplicate(
     ctx.throw(404)
   }
 
-  const duplicatedApp = await createDuplicatedApp(workspaceApp)
-  await duplicateScreens(
-    workspaceApp._id as string,
-    duplicatedApp._id as string
-  )
+  const duplicatedApp = await sdk.workspaceApps.duplicate(workspaceApp)
 
   ctx.message = `App ${workspaceApp.name} duplicated successfully.`
   ctx.body = { workspaceApp: toWorkspaceAppResponse(duplicatedApp) }
